requirements.txt
requests==2.32.3
python-dotenv==1.1.0
geocoder==1.38.1
pytz==2025.2
fastapi==0.115.12
pydantic-settings==2.9.1
uvicorn==0.34.3
pytest==8.4.0
httpx==0.28.1
pytest-asyncio==1.0.0
pyyaml==6.0.2